@@ -20,16 +20,20 @@
   "prefer-stable": true,
   "config": {
     "platform": {
-      "php": "7.1.3"
+      "php": "8.2"
     },
     "vendor-dir": "lib"
   },
   "require": {
-    "php": ">=7.1.3",
-    "jtl/connector": "^3.2.0",
+    "php": ">=8.1",
+    "jtl/connector": "^5.2.1",
     "symfony/yaml": "^3.4",
     "ext-json": "*",
-    "phing/phing": "^2.16"
+    "phing/phing": "^2.16",
+    "psr/cache": "^2",
+    "monolog/monolog": "^2",
+    "psr/log": "^2",
+    "symfony/http-foundation": "^4.4"
   },
   "autoload": {
     "psr-4": {
@@ -41,7 +45,6 @@
     }
   },
   "require-dev": {
-    "jtl/connector-integration-tests": "dev-develop",
     "friendsofphp/php-cs-fixer": "^2.16",
     "symfony/polyfill-php80": "<=1.18.1"
   }
