@@ -4,7 +4,7 @@ namespace jtl\Connector\Presta\Mapper;
 
 class CategoryInvisibility extends BaseMapper
 {
-    protected $pull = [
+    protected array $pull = [
         'categoryId'      => 'id_category',
         'customerGroupId' => 'id_group'
     ];
